Move internal/kptfile to pkg/kptfile to make it vendorable
