+++ internal/cmdget/cmdget_test.go
@@ -22,8 +22,8 @@ import (
 	"testing"
 
 	"github.com/GoogleContainerTools/kpt/internal/cmdget"
-	"github.com/GoogleContainerTools/kpt/internal/kptfile"
 	"github.com/GoogleContainerTools/kpt/internal/testutil"
+	"github.com/GoogleContainerTools/kpt/pkg/kptfile"
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/assert"
 	"sigs.k8s.io/kustomize/kyaml/yaml"
